backend/requirements.txt
fastapi
uvicorn
sqlalchemy
requests
geopy
jinja2
python-multipart
python-dotenv
pydantic
